+++ src/materials/phong.cpp
@@ -9,12 +9,12 @@
 #include <cmath>
 
 const rgbColor phongBrdf::sampleF(const float& u0, const float& u1, const vec3& wo, vec3& wi, float& pd) const{
-    const float sinAlpha = sqrtf(1.f - powf(u0, 2.f / (float)(n+1)));
+    const float sinAlpha = sqrt(1.f - pow(u0, 2.f / (float)(n+1)));
     const float phi = TWOPI * u1;
 
-    wi.x = sinAlpha * cosf(phi);
-    wi.y = powf(u0, (1.f / (float)(n+1)));
-    wi.z = sinAlpha * sinf(phi);
+    wi.x = sinAlpha * cos(phi);
+    wi.y = pow(u0, (1.f / (float)(n+1)));
+    wi.z = sinAlpha * sin(phi);
 
     if(wi.y < 0){
         pd = 0.f;
@@ -30,7 +30,7 @@ const rgbColor phongBrdf::f(const vec3& wo, const vec3& wi) const{
     if(cosAlpha < 0){
         return rgbColor(0.f);
     }else{
-        return ks * (float)(n+2) * INVTWOPI * powf(cosAlpha, n);
+        return ks * (float)(n+2) * INVTWOPI * pow(cosAlpha, n);
     }
 }
 
@@ -39,6 +39,6 @@ float phongBrdf::pdf(const vec3& wo, const vec3& wi) const {
     if(cosAlpha < 0){
         return 0.f;
     }else{
-        return ((float)(n+1) * INVTWOPI) * powf(cosAlpha, n);
+        return ((float)(n+1) * INVTWOPI) * pow(cosAlpha, n);
     }
 }
